:recycle: Service Transaction 전략 수정 (#151)
