+++ src/main/java/com/smunity/api/domain/account/service/UserService.java
@@ -1,4 +1,4 @@
-package com.smunity.api.domain.account.service.impl;
+package com.smunity.api.domain.account.service;
 
 import com.smunity.api.domain.account.repository.UserRepository;
 import com.smunity.api.global.error.exception.RestException;
@@ -7,10 +7,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-@RequiredArgsConstructor
 @Service
-public class UserDetailsServiceImpl implements UserDetailsService {
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class UserService implements UserDetailsService {
     private final UserRepository userRepository;
 
     @Override
